Revert "[Box] Deprecate css prop in favor of sx (#23480)"
This reverts commit c3de8e5ef7b96c49b4833ef55e7fbf2cf09a4592.
