+++ docs/src/pages/system/basics/CssProp.js
@@ -2,11 +2,11 @@ import React from 'react';
 import styled, { ThemeProvider } from 'styled-components';
 import NoSsr from '@material-ui/core/NoSsr';
 import { createMuiTheme } from '@material-ui/core/styles';
-import { compose, spacing, palette, styleFunctionSx } from '@material-ui/system';
+import { compose, spacing, palette, css } from '@material-ui/system';
 
-const styleFunction = styleFunctionSx(compose(spacing, palette));
-
-const Box = styled.div(styleFunction);
+const Box = styled.div`
+  ${css(compose(spacing, palette))}
+`;
 
 const theme = createMuiTheme();
 
@@ -14,7 +14,7 @@ export default function CssProp() {
   return (
     <NoSsr>
       <ThemeProvider theme={theme}>
-        <Box color="white" sx={{ bgcolor: 'palevioletred', p: 1, textTransform: 'uppercase' }}>
+        <Box color="white" css={{ bgcolor: 'palevioletred', p: 1, textTransform: 'uppercase' }}>
           CssProp
         </Box>
       </ThemeProvider>
