+++ docs/src/pages/system/flexbox/AlignItems.js
@@ -10,7 +10,7 @@ export default function AlignItems() {
         p={1}
         m={1}
         bgcolor="background.paper"
-        sx={{ height: 100 }}
+        css={{ height: 100 }}
       >
         <Box p={1} bgcolor="grey.300">
           Item 1
@@ -28,7 +28,7 @@ export default function AlignItems() {
         p={1}
         m={1}
         bgcolor="background.paper"
-        sx={{ height: 100 }}
+        css={{ height: 100 }}
       >
         <Box p={1} bgcolor="grey.300">
           Item 1
@@ -46,7 +46,7 @@ export default function AlignItems() {
         p={1}
         m={1}
         bgcolor="background.paper"
-        sx={{ height: 100 }}
+        css={{ height: 100 }}
       >
         <Box p={1} bgcolor="grey.300">
           Item 1
