+++ docs/src/pages/system/flexbox/FlexWrap.js
@@ -10,7 +10,7 @@ export default function FlexWrap() {
         p={1}
         m={1}
         bgcolor="background.paper"
-        sx={{ maxWidth: 300 }}
+        css={{ maxWidth: 300 }}
       >
         <Box p={1} bgcolor="grey.300">
           Item 1
@@ -37,7 +37,7 @@ export default function FlexWrap() {
         p={1}
         m={1}
         bgcolor="background.paper"
-        sx={{ maxWidth: 300 }}
+        css={{ maxWidth: 300 }}
       >
         <Box p={1} bgcolor="grey.300">
           Item 1
