+++ packages/material-ui-system/src/index.d.ts
@@ -53,19 +53,10 @@ export type ComposedStyleFunction<T extends Array<StyleFunction<any>>> = StyleFu
 >;
 export function compose<T extends Array<StyleFunction<any>>>(...args: T): ComposedStyleFunction<T>;
 
-// styleFunctionSx.js
-
-/**
- * @deprecated
- * The css style function is deprecated. Use the styleFunctionSx instead.
- */
+// css.js
 export function css<Props>(
   styleFunction: StyleFunction<Props>
-): StyleFunction<Props & { css?: Omit<Props, 'theme'>; sx?: Omit<Props, 'theme'> }>;
-
-export function styleFunctionSx<Props>(
-  styleFunction: StyleFunction<Props>
-): StyleFunction<Props & { sx?: Omit<Props, 'theme'>; css?: Omit<Props, 'theme'> }>;
+): StyleFunction<Props & { css: Omit<Props, 'theme'> }>;
 
 export const display: SimpleStyleFunction<
   'display' | 'displayPrint' | 'overflow' | 'textOverflow' | 'visibility' | 'whiteSpace'
