+++ packages/material-ui-system/src/index.js
@@ -2,8 +2,7 @@ export { default as borders } from './borders';
 export * from './borders';
 export { default as breakpoints } from './breakpoints';
 export { default as compose } from './compose';
-export { default as styleFunctionSx } from './styleFunctionSx';
-export * from './styleFunctionSx';
+export { default as css } from './css';
 export { default as display } from './display';
 export { default as flexbox } from './flexbox';
 export * from './flexbox';
