+++ packages/material-ui/src/Box/Box.test.js
@@ -1,8 +1,6 @@
 import * as React from 'react';
-import PropTypes from 'prop-types';
 import { expect } from 'chai';
 import { createClientRender } from 'test/utils/createClientRender';
-import consoleErrorMock from 'test/utils/consoleErrorMock';
 import createMount from 'test/utils/createMount';
 import describeConformance from '@material-ui/core/test-utils/describeConformance';
 import Box from './Box';
@@ -10,14 +8,6 @@ import Box from './Box';
 describe('<Box />', () => {
   const mount = createMount();
   const render = createClientRender();
-
-  beforeEach(() => {
-    consoleErrorMock.spy();
-  });
-
-  afterEach(() => {
-    consoleErrorMock.reset();
-  });
 
   describeConformance(<Box />, () => ({
     mount,
@@ -56,14 +46,5 @@ describe('<Box />', () => {
     expect(element.getAttribute('color')).to.equal(null);
     expect(element.getAttribute('font-family')).to.equal(null);
     expect(element.getAttribute('font-size')).to.equal(null);
-  });
-
-  it('warns if the css prop is used ', () => {
-    PropTypes.checkPropTypes(Box.propTypes, { css: { m: 1, p: 1 } }, 'props', 'MockedBox');
-
-    expect(consoleErrorMock.callCount()).to.equal(1);
-    expect(consoleErrorMock.messages()[0]).to.include(
-      'Material-UI: The `css` prop is deprecated, please use the `sx` prop instead.',
-    );
   });
 });
